make the include macro check better
The include macro is now checked against only the beginning of the line,
and comments in the asm file can say "INCLUDE \"" if they need to. In
addition, the preprocessor now supports INCLUDE lines that are preceded
by a tab character. However, if this included file has multiple lines,
the output will not be automatically aligned with the initial tab. This
might cause problems with rgbasm.