make a vagrant box to make things easier
Maybe a vagrant box will be less work for users. At the moment the
Vagrantfile is just pointing to a remote url, but in the future it
should contain the actual scripts to construct the box so that the box
can be updated as the dependencies in the repository change.