Use connection macros instead of guessed values in second map headers.
Some of the under-the-hood connection values are offset by an arbitrary amount. This could be because of a graphical connection editor allowing a shift in the strip destination pointer, as well as the standard alignment value. There is no visible difference, so it can be confusing when two unique sets of values can produce the same result.

Reversing this invisible offset was a major sticking point back when connections were dumped, but it was straightforward this time around. Much thanks to huderlem for articulating this quirk in pokered's connection macros.