Add initial linkerscript
This linkerscript will sequentially allocate all the sections that were previously not fixed. This allows such sections to remain in place relative to each other, regardless of if their size is changed (e.g. by mods).

Only the minimum necessary ORGs are included. The need for these may be removed in the future by including more sections in the script.

As part of this change, I had to rename the “Audio” section in WRAM0 to “WRAM Audio“ (similar to some SRAM sections), as there were conflicts with the identically-named ROMX section.