Remove all unused functions
Added two scripts in tools/. One is a python script responsible for
extracting unused symbols from .o files, the other a shell script
responsible for preparing the source and filtering the output as
appropriate.

Using this, I've removed all functions and data belonging to unused symbols.
This is just a quick glance into what is going to be an even bigger
cleanup, but for now I haven't modified any existing functions to be
able to remove even more. This will be touched upon in later commits.