Add MobileError-related labels
This function is called from a few different places in the mobile code,
including from one special that is only executed when the
BattleTowerRoomMenu special returns with a non-zero code.

Since that special causes a blue "error" screen to appear, that looks
the exact same as the one you'd get when trying to connect from the
mobile trade corner/battle colosseum in a Pokémon Center's top floor,
and this function contains a bunch of japanese strings in a jumptable, I
think it's reasonable to assume it should be called
"DisplayMobileError".