Removed all mobile-related code
This pretty huge commit removes everything mobile and mobile-related in
the ROM. This has freed up approximately 5% of the space, and cleaned up
a bunch of otherwise useless files. There's still leftovers that will be
fixed in a further commit.

Not a lot was optimized. I just cleaned up a lot of tables that referenced
mobile-related code and cleaned up from there. The only thing I settled
to optimize was misc/mobile_46.asm, which referenced a ton of
otherwise-unused wram addresses.

An overview of the changes:
- Cleaned up battle tower functions to remove mobile functionality
- Removed mobile-only scripts in maps
- Removed battle and trade-room maps, because they only referenced mobile code.
- Removed mobile and battle studium entries in the main menu
- Removed card exchange through mystery gift code

And, of course, I removed everything left unreferenced.