Remove Celebi event
It's kind of a weird case, this one.
On one hand, this event is pretty darn cool, and could easyly be used as-is
by ROM hackers that want it.
On the other hand, there's no way to trigger it without hacking or
glitching, so it might as well not exist, and it'll be useless in most
hacks.
Since I could squeeze a few more bytes from the ROM by removing it, I
decided I might as well. If someone feels the need to include it, a
simple `git revert` of this commit should do.