Added support for building a Virtual Console patch
This commit adds a program that is to aid in "fixing" the patches
provided by nintendo in their Virtual Console releases, as well as
potentially expanding upon them.

In doing so, I've disassembled the patch found in the Pokémon Crystal
Virtual Console release.

Of note is that this Virtual Console release has been advertised as
being able to unlock the Celebi event at the Goldenrod City Pokémon Center,
yet no patches to achieve this in the code actually exist. The emulator
handles this by editing the save file, which is fatal for a lot of ROM
hacks.