Split everything into separate `.o` files
This commit is the start of what I call "the `.o` split".
Sadly, this split has made the entire build system _magnitudes_ slower.
As such, I will try to make it less slow, and hopefully improve the
incremental build times, and later the build time overall.

But, if it doesn't work out, this can always be reverted.