Add travis support for checking unused globals
This will make sure only necessary symbols are ever exported, except for
everything in the ram sections, of course, which doesn't make sense to
not be exported.

engine/predef.asm and engine/events/specials.asm are ignored because the
arrays generate a bunch of symbols that are unused, but we can't really
not export some of them since we use a macro.