Include some constants inside of maps
floating_constants.inc was moved out of maps_common.inc because it's not
something that is necessary for new maps - it's supposed to be the
exception, not the norm.

Since there's only four maps that use BattleTowerAction, and the
battle_tower_constants.inc aren't otherwise used anywhere, they're now
only included where required.