Remove all text_far comments
Rationale: These comments are inconsistently used, don't often display
the full extent of the string (missing characters and whatnot), and
are about as useful to hackers as the address comments were. They're
also kind of a hindrance for i18n support, as it'd require overlaying
some files purely for the comments.

If we actually want them, we should have some tool to make sure they're
consistently used, and easy to update.