Correctly fix faulty Full Heal of the AI
The fix for "AI use of Full Heal does not cure confusion status" worked to cure confusion, but it set `a` to 0 when `jp PrintText_UsedItemOn_AND_AIUpdateHUD` needs it to be set to `FULL_HEAL`, as it is just before. It also added a useless `ld a, [wCurPartyItem]` since `PrintText_UsedItemOn_AND_AIUpdateHUD` already does it.