Use hex instead of characters for codepoints
When referencing a codepoint as a codepoint, ensuring that a
hexadecimal value is used instead of the character that usually
occupies that codepoint. In all such cases, the original character has
been left as a comment, so this information is not lost.

Additionally, in correct_nick_errors.asm and mobile_5f.asm,
added a comment that highlights the differences from the other function
that serves a similar purpose.