Rename MON_ID to MON_TRAINER_ID
The constant name MON_ID implies it relates to the ID number of the
Pokémon itself. However, it actually refers to the Trainer ID number of
the Pokémon's Original Trainer. Renaming it to MON_TRAINER_ID makes it
substantially more clear what its purpose is.

The original definition of MON_ID is preserved in macros/legacy.asm